# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/projects/me": {
      "get": {
        "operationId": "ProjectsController_listProjects",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": { "type": "number" }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": { "type": "number" }
          },
          {
            "name": "keyword",
            "required": false,
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectPaginateEntity"
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Projects"]
      }
    },
    "/projects/create": {
      "post": {
        "operationId": "ProjectsController_createProject",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateProjectDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProjectEntity" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Projects"]
      }
    },
    "/projects/me/{id}": {
      "get": {
        "operationId": "ProjectsController_getProject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProjectEntity" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Projects"]
      },
      "patch": {
        "operationId": "ProjectsController_updateProject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateProjectDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Project updated successfully.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProjectEntity" }
              }
            }
          },
          "401": { "description": "" },
          "404": { "description": "Project not found." }
        },
        "security": [{ "bearer": [] }, { "bearer": [] }],
        "summary": "Update an existing project",
        "tags": ["Projects"]
      },
      "delete": {
        "operationId": "ProjectsController_deleteProject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Project deleted successfully." },
          "401": { "description": "" },
          "404": { "description": "Project not found." }
        },
        "security": [{ "bearer": [] }, { "bearer": [] }],
        "summary": "Delete a project",
        "tags": ["Projects"]
      }
    },
    "/projects/{id}/roles": {
      "get": {
        "operationId": "ProjectsController_getRoles",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/RoleEntity" }
                }
              }
            }
          },
          "401": { "description": "" }
        },
        "security": [{ "bearer": [] }, { "bearer": [] }],
        "tags": ["Projects"]
      }
    },
    "/projects/{projectid}/roles": {
      "post": {
        "operationId": "ProjectsController_createRole",
        "parameters": [
          {
            "name": "projectid",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateRoleDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProjectEntity" }
              }
            }
          },
          "401": { "description": "" }
        },
        "security": [{ "bearer": [] }, { "bearer": [] }],
        "tags": ["Projects"]
      }
    },
    "/projects/{projectid}/roles/{roleid}": {
      "delete": {
        "operationId": "ProjectsController_deleteRole",
        "parameters": [
          {
            "name": "projectid",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "roleid",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/GeneralResult" }
              }
            }
          },
          "401": { "description": "" }
        },
        "security": [{ "bearer": [] }, { "bearer": [] }],
        "tags": ["Projects"]
      }
    },
    "/projects/{projectid}/members": {
      "get": {
        "operationId": "ProjectsController_getMembers",
        "parameters": [
          {
            "name": "projectid",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberPaginateEntity"
                }
              }
            }
          },
          "401": { "description": "" }
        },
        "security": [{ "bearer": [] }, { "bearer": [] }],
        "tags": ["Projects"]
      }
    },
    "/projects/{projectid}/members/{userid}": {
      "delete": {
        "operationId": "ProjectsController_deleteMember",
        "parameters": [
          {
            "name": "projectid",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "userid",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/GeneralResult" }
              }
            }
          },
          "401": { "description": "" }
        },
        "security": [{ "bearer": [] }, { "bearer": [] }],
        "tags": ["Projects"]
      }
    },
    "/projects/{projectid}/invite": {
      "get": {
        "operationId": "ProjectsController_getInvitationLink",
        "parameters": [
          {
            "name": "projectid",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "email",
            "required": true,
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "roleid",
            "required": true,
            "in": "query",
            "schema": { "type": "number" }
          }
        ],
        "responses": { "401": { "description": "" } },
        "security": [{ "bearer": [] }, { "bearer": [] }],
        "tags": ["Projects"]
      }
    },
    "/projects/{projectid}/invite/verify": {
      "get": {
        "operationId": "ProjectsController_verifyMember",
        "parameters": [
          {
            "name": "projectid",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          },
          {
            "name": "token",
            "required": true,
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/GeneralResult" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "tags": ["Projects"]
      }
    },
    "/projects/{projectid}/status": {
      "post": {
        "operationId": "ProjectsController_createStatus",
        "parameters": [
          {
            "name": "projectid",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectStatusDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The status has been successfully created.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/StatusEntity" }
              }
            }
          },
          "400": { "description": "Invalid input data." }
        },
        "security": [{ "bearer": [] }],
        "summary": "Create a new status",
        "tags": ["Projects"]
      },
      "get": {
        "operationId": "ProjectsController_findAllStatus",
        "parameters": [
          {
            "name": "projectid",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "List of statuses retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/StatusEntity" }
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Get all statuses",
        "tags": ["Projects"]
      }
    },
...
    "/projects/{projectid}/category": {
      "post": {
        "operationId": "ProjectsController_createCategory",
        "parameters": [
          {
            "name": "projectid",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateCategoryDto" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The category has been successfully created.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CategoryEntity" }
              }
            }
          },
          "400": { "description": "Invalid input data." }
        },
        "security": [{ "bearer": [] }],
        "summary": "Create a new category",
        "tags": ["Projects"]
      },
      "get": {
        "operationId": "ProjectsController_findAllCategory",
        "parameters": [
          {
            "name": "projectid",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "List of categories retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/CategoryEntity" }
                }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Get all categories",
        "tags": ["Projects"]
      }
    },
...
  },
  "info": {
    "title": "THESIS",
    "description": "The API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" }
    },
    "schemas": {
...
      "RoleEntity": {
        "type": "object",
        "properties": {
          "id": { "type": "number" },
          "category": {
            "type": "string",
            "enum": ["MEMBER", "GUEST", "OWNER"]
          },
          "name": { "type": "string" },
          "projectId": { "type": "number" }
        },
        "required": ["id", "category", "name", "projectId"]
      },
...
      "ProjectEntity": {
        "type": "object",
        "properties": {
          "id": { "type": "number" },
          "name": { "type": "string" },
          "projectThumbnail": { "type": "string" },
          "description": { "type": "string" },
          "createdAt": { "format": "date-time", "type": "string" },
          "updatedAt": { "format": "date-time", "type": "string" },
          "userRole": { "$ref": "#/components/schemas/UserRoleEntity" },
          "githubOrgName": { "type": "string" },
          "GithubRepo": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/GithubRepoEntity" }
          }
        },
        "required": [
          "id",
          "name",
          "projectThumbnail",
          "description",
          "createdAt",
          "updatedAt"
        ]
      },
      "ProjectPaginateEntity": {
        "type": "object",
        "properties": {
          "total": { "type": "number" },
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ProjectEntity" }
          }
        },
        "required": ["total", "items"]
      },
...
      "CreateProjectDto": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "description": { "type": "string" },
          "projectThumbnail": { "type": "string" },
          "projectDomain": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/CreateProjectDomainDto" }
          }
        },
        "required": ["name", "description"]
      },
...
    }
  },
  "security": [{ "bearer": [] }]
}